+++ controllers/GalleryController.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace app\controllers;
+
+use Yii;
+use yii\web\Controller;
+
+class GalleryController extends Controller
+{
+
+    public function actions()
+    {
+        return [
+            'error' => [
+                'class' => 'yii\web\ErrorAction',
+            ],
+        ];
+    }
+
+    public function actionIndex()
+    {
+        return $this->render('index');
+    }
+}
