+++ views/gallery/index.php
