+++ views/layouts/main.php
@@ -26,7 +26,7 @@ AppAsset::register($this);
     <div class="wrap">
         <?php
             NavBar::begin([
-                'brandLabel' => 'My Company',
+                'brandLabel' => Yii::$app->name,
                 'brandUrl' => Yii::$app->homeUrl,
                 'options' => [
                     'class' => 'navbar-inverse navbar-fixed-top',
@@ -37,7 +37,7 @@ AppAsset::register($this);
                 echo Nav::widget([
                     'options' => ['class' => 'navbar-nav navbar-right'],
                     'items' => [
-                        ['label' => Yii::t('pg.main', 'Home'), 'url' => ['/site/index']],
+                        ['label' => Yii::t('pg.main', 'Home'), 'url' => ['/gallery/index']],
                         ['label' => Yii::t('pg.main', 'Sign Up'), 'url' => ['/auth/default/signup']],
                         ['label' => Yii::t('pg.main', 'Sign In'), 'url' => ['/auth/default/login']]
                     ],
@@ -46,9 +46,9 @@ AppAsset::register($this);
                 echo Nav::widget([
                     'options' => ['class' => 'navbar-nav navbar-right'],
                     'items' => [
-                        ['label' => Yii::t('pg.main', 'Home'), 'url' => ['/site/index']],
+                        ['label' => Yii::t('pg.main', 'Home'), 'url' => ['/gallery/index']],
                         ['label' => Yii::t('pg.main', 'Sign Out'), ' (' . Yii::$app->user->identity->username . ')',
-                            'url' => ['/site/logout'],
+                            'url' => ['/auth/default/logout'],
                             'linkOptions' => ['data-method' => 'post']],
                     ],
                 ]);
@@ -63,13 +63,6 @@ AppAsset::register($this);
             <?= $content ?>
         </div>
     </div>
-
-    <footer class="footer">
-        <div class="container">
-            <p class="pull-left">&copy; My Company <?= date('Y') ?></p>
-            <p class="pull-right"><?= Yii::powered() ?></p>
-        </div>
-    </footer>
 
 <?php $this->endBody() ?>
 </body>
